Conflicts:
	domain/src/main/java/org/fao/geonet/domain/MetadataIdentifierTemplate.java
	domain/src/main/java/org/fao/geonet/repository/specification/MetadataIdentifierTemplateSpecs.java
	domain/src/test/java/org/fao/geonet/repository/specification/MetadataIdentifierTemplateSpecsTest.java
	services/src/main/java/org/fao/geonet/services/metadata/Create.java
	services/src/main/java/org/fao/geonet/services/metadata/MetadataIdentifierTemplatesController.java
	web-ui/src/main/resources/catalog/components/catalog/CatalogService.js
	web-ui/src/main/resources/catalog/js/admin/MetadataIdentifierTemplatesController.js
	web-ui/src/main/resources/catalog/js/edit/NewMetadataController.js
	web-ui/src/main/resources/catalog/locales/du-editor.json
	web-ui/src/main/resources/catalog/locales/en-admin.json
	web-ui/src/main/resources/catalog/locales/es-admin.json
	web-ui/src/main/resources/catalog/locales/es-editor.json
	web-ui/src/main/resources/catalog/locales/fr-admin.json
	web-ui/src/main/resources/catalog/style/ellipsis.less
	web/src/main/webapp/WEB-INF/classes/setup/sql/data/data-db-default.sql
	web/src/main/webapp/WEB-INF/config-security/config-security-mapping.xml

+++ services/src/main/java/org/fao/geonet/services/resources/Get.java
@@ -56,6 +56,12 @@ public class Get extends NotInReadOnlyModeService {
 
         // delete the file
         Path file = Lib.resource.getDir(context, access, id).resolve(filename);
+        
+        if(filename.contains("..")
+                || filename.startsWith("://", 1) 
+                || filename.startsWith("/")) {
+            throw new SecurityException("Wrong filename");
+        }
 
         Files.deleteIfExists(file);
